clean_ups.py modifications

+++ server/labml_app/clean_ups.py
@@ -34,16 +34,16 @@ def clean_float_project() -> None:
             r = run_key.load()
             s = r.status.load()
 
-            if (time.time() - 86400) > s.last_updated_time:
+            if not r.is_claimed and (time.time() - 86400) > s.last_updated_time:
                 run.delete(run_uuid)
                 delete_run_list.append(run_uuid)
+            elif (time.time() - 86400) > s.last_updated_time:
+                delete_run_list.append(run_uuid)
         except TypeError:
-            logger.error(f'error while deleting the run {run_uuid}')
-            delete_run_list.append(run_uuid)
+            logger.error(f'error while deleting the session {run_uuid}')
 
     for run_uuid in delete_run_list:
         p.runs.pop(run_uuid)
-
     p.save()
 
     logger.info('......Done.........')
@@ -55,16 +55,16 @@ def clean_float_project() -> None:
             ss = session_key.load()
             s = ss.status.load()
 
-            if (time.time() - 86400) > s.last_updated_time:
+            if not ss.is_claimed and (time.time() - 86400) > s.last_updated_time:
                 session.delete(session_uuid)
+                delete_session_list.append(session_uuid)
+            elif (time.time() - 86400) > s.last_updated_time:
                 delete_session_list.append(session_uuid)
         except TypeError:
             logger.error(f'error while deleting the session {session_uuid}')
-            delete_run_list.append(session_uuid)
 
     for session_uuid in delete_session_list:
         p.sessions.pop(session_uuid)
-
     p.save()
 
     logger.info('......Done.........')
@@ -75,6 +75,10 @@ def move_to_samples():
     p = project.get_project(settings.SAMPLES_PROJECT_TOKEN)
     for run_uuid in block_uuids.delete_run_uuids:
         r = run.get_run(run_uuid)
+
+        if r.owner == 'samples':
+            continue
+
         if r and r.owner != 'samples':
             r.owner = 'samples'
             p.add_run(run_uuid)
